Ajustes e testes no serviço

+++ BackgroundServices/Scripts/MySQL/Criação Cenário MySQL.sql
@@ -15,14 +15,13 @@ create table tb_depusr(
 	, constraint cf_depusr_01 foreign key (usuar_cod) REFERENCES tb_usuar(usuar_cod)
 );
 
-
 create table tb_depusrSINCRO(	
 	emp_cod	numeric not null
 	, depend_cod numeric not null
 	, usuar_cod	numeric not null
     , SCHAVEINTEGRACAO varchar(50) null
     , SMETODO varchar(10) not null
-    , SQTDETENTATIVA numeric not null
+    , SQTDETENTATIVA int not null
 	, SSTATUSINTEGRACAO varchar(10) not null
 	, SDATAINTEGRACAO datetime not null
 );
